+++ lib/nretnil/compendium-api/callback.rb
@@ -15,8 +15,14 @@ module Nretnil
         @session = s
       end
 
-      def fire(name, arguments)
+      def fire(name, arguments = {}, network_id = nil, schedule_at = nil)
           query = { :EventName => name, :EventArgs => arguments.to_json }
+          if network_id
+            query[:NetworkId] = network_id
+          end
+          if schedule_at
+            query[:ScheduleAt] = schedule_at
+          end
           response = @session.post( '/app/callback/fire', query )
       end